Move the Application.start function into its own module
